fix: resolve Phase 5 lint violations for PERF, SIM, RUF, C4 rules
This commit fixes remaining violations identified after enabling Phase 5
linter rules:

Code Fixes:
- SIM101: Merge multiple isinstance() calls into single call
- SIM103: Return condition directly instead of if-true-return-false
- SIM210/SIM211: Remove unnecessary True/False in ternary expressions
- C401/C403/C405: Rewrite list/generator to set comprehension
- C414/C419: Remove unnecessary list() in sorted()/all()
- RUF043: Use raw strings for regex patterns in pytest.raises match
- RUF046: Remove unnecessary int() call on already-int values
- FURB187: Added to ignore (Sequence cannot be reversed in place)

Configuration Updates (pyproject.toml):
- Added FURB187 to ignore list (Sequence types cannot use .reverse())
- Added RUF003 to ignore list (intentional Unicode chars in comments)
- Added RUF009 to ignore list (cast() in dataclass defaults is intentional)
- Added per-file-ignore for LOG015 in otel.py (root logger usage intentional)

+++ src/ai/backend/runner/hash_phrase.py
@@ -43,7 +43,7 @@ def hash_phrase(
 
     dict_len = len(dictionary)
     entropy_per_word = math.log2(dict_len)
-    num_words = int(math.ceil(minimum_entropy / entropy_per_word))
+    num_words = math.ceil(minimum_entropy / entropy_per_word)
 
     # Hash the data and convert to a big integer (converts as Big Endian)
     hash = hashfunc(data)
